🚚 move: main, post 파일 분리리
