+++ routes.jsx
@@ -1,17 +1,17 @@
 import MyPage from "@pages/myPage/MyPage";
 import Edit from "@pages/myPage/Edit";
 import Likes from "@pages/myPage/Likes";
-import PRWrite from "@pages/pr/PRWrite";
+import PRWrite from "@pages/main/post/PRWrite";
 import SignIn from "@pages/user/SignIn";
 import SignUp from "@pages/user/SignUp";
 import Terms from "@pages/user/Terms";
 import { createBrowserRouter, Navigate } from "react-router-dom";
 import MyReviews from "@pages/myPage/MyReviews";
 import Layout from "@components/layout";
-import PostDetail from "@pages/main/PostDetail";
+import PostDetail from "@pages/main/post/PostDetail";
 import PostList from "@pages/main/PostList";
-import PostWrite from "@pages/main/PostWrite";
-import PostEdit from "@pages/main/PostEdit";
+import PostWrite from "@pages/main/post/PostWrite";
+import PostEdit from "@pages/main/post/PostEdit";
 import ReviewList from "@pages/history/Layout";
 import Profile from "@pages/user/Profile";
 import Error from "@pages/Error";
